add route for templates
